Renamed ValueController and cleaned up some not so nice code
